commit: create report

+++ src/main/java/com/nwchat/сontroller/OrderController.java
@@ -1,8 +1,7 @@
 package com.nwchat.сontroller;
 
 import com.nwchat.entity.OrderEntity;
-import com.nwchat.entity.UserEntity;
-import com.nwchat.model.Doc;
+import com.nwchat.entity.ReportEntity;
 import com.nwchat.repository.OrderRepository;
 import com.nwchat.repository.UserRepository;
 import com.nwchat.service.UserService;
@@ -34,19 +33,47 @@ public class OrderController {
 	@RequestMapping(value = "/", method = RequestMethod.GET)
     public ModelAndView index() {
         ModelAndView model = new ModelAndView();
-        Iterable<OrderEntity> listOrder =  orderRepository.findAll();
-        model.addObject("orderList", listOrder);
-        model.setViewName("order/index");
+		Integer roleId = userService.getAuthenticationUser().getRoleId();
+        if (roleId==1) {
+			Iterable<OrderEntity> listOrder = orderRepository.findAll();
+			model.addObject("orderList", listOrder);
+			model.addObject("userRole", roleId);
+			model.setViewName("order/index");
+		}
+        else if(roleId==2){
+        	List<OrderEntity> listOrders = orderRepository.findAllByManagerIdEquals(userService.getAuthenticationUser().getId());
+        	model.addObject("orderList",listOrders);
+			model.addObject("userRole", roleId);
+			model.setViewName("order/index");
+		}
         return model;
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
 	public ModelAndView show(@PathVariable(value = "id") int id) {
 		ModelAndView model = new ModelAndView();
-
 		OrderEntity orderEntity = orderRepository.findById(id).get();
-        model.addObject("order", orderEntity);
+		Integer roleId = userService.getAuthenticationUser().getRoleId();
+		model.addObject("userRole", roleId);
+		model.addObject("order", orderEntity);
 		model.setViewName("order/show");
+
+		return model;
+	}
+	@RequestMapping(value = "/{id}/send",method = RequestMethod.GET)
+	public ModelAndView send(@PathVariable int id ){
+		ModelAndView model = new ModelAndView();
+		OrderEntity orderEntity = orderRepository.findById(id).get();
+		Integer roleId = userService.getAuthenticationUser().getRoleId();
+		if (roleId ==1) {
+			orderEntity.setState(1);
+			orderEntity = orderRepository.save(orderEntity);
+			model.addObject("order",orderEntity);
+			model.setViewName("redirect:/order/");
+		}
+		else {
+			model.setViewName("redirect:/order/");
+		}
 		return model;
 	}
 
@@ -106,8 +133,15 @@ public class OrderController {
 	@RequestMapping(value = "/{id}/delete")
 	public ModelAndView delete(@PathVariable int id) {
 		ModelAndView model = new ModelAndView();
-		orderRepository.deleteById(id);
-		model.setViewName("redirect:/order/");
+		Integer roleId = userService.getAuthenticationUser().getRoleId();
+		OrderEntity orderEntity = orderRepository.findById(id).get();
+		Integer state = orderEntity.getState();
+		if ((roleId ==1) && (state==0)) {
+			orderRepository.deleteById(id);
+			model.setViewName("redirect:/order/");
+		} else {
+			model.setViewName("redirect:/order/");
+		}
 		return model;
 	}
